Crate arrow_ipc

Support for the Arrow IPC Format

The Arrow IPC format defines how to read and write RecordBatches to/from a file or stream of bytes. This format can be used to serialize and deserialize data to files and over the network.

There are two variants of the IPC format:

IPC Streaming Format: Supports streaming data sources, implemented by StreamReader and StreamWriter

IPC File Format: Supports random access, implemented by FileReader and FileWriter.

See the reader and writer modules for more information.

Re-exports

pub use self::gen::File::*;

pub use self::gen::Message::*;

pub use self::gen::Schema::*;

pub use self::gen::SparseTensor::*;

pub use self::gen::Tensor::*;

Modules

compression 🔒

convert

Utilities for converting between IPC types and native Arrow types

gen

Generated code

reader

Arrow IPC File and Stream Readers

writer

Arrow IPC File and Stream Writers

Constants

ARROW_MAGIC 🔒

CONTINUATION_MARKER 🔒