arrow_ipc::gen::File

Function root_as_footer

pub fn root_as_footer(buf: &[u8]) -> Result<Footer<'_>, InvalidFlatbuffer>

Verifies that a buffer of bytes contains a Footer and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_footer_unchecked.