Module arrow_ipc::reader

Arrow IPC File and Stream Readers

The FileReader and StreamReader have similar interfaces, however the FileReader expects a reader that supports Seeking

Structs

FileDecoder

A low-level, push-based interface for reading an IPC file

FileReader

Arrow File reader

FileReaderBuilder

Build an Arrow FileReader with custom options.

StreamReader

Arrow Stream reader

Functions

read_dictionary

Read the dictionary from the buffer and provided metadata, updating the dictionaries_by_id with the resulting dictionary

read_footer_length

Read the footer length from the last 10 bytes of an Arrow IPC file

read_record_batch

Creates a record batch from binary data using the crate::RecordBatch indexes and the Schema