arrow_ipc::reader

Struct MessageReader

struct MessageReader<R> {
    reader: R,
    buf: Vec<u8>,
}

A low-level construct that reads Message::Messages from a reader while re-using a buffer for metadata. This is composed into StreamReader.

Fields
reader: Rbuf: Vec<u8>
Implementations

impl<R: Read> MessageReader<R>

fn new(reader: R) -> Self

fn maybe_next( &mut self, ) -> Result<Option<(Message<'_>, MutableBuffer)>, ArrowError>

Reads the entire next message from the underlying reader which includes the metadata length, the metadata, and the body.

Returns

Ok(None) if the the reader signals the end of stream with EOF on the first read

Err(_) if the reader returns an error other than on the first read, or if the metadata length is invalid

Ok(Some(_)) with the Message and buffer containiner the body bytes otherwise.

fn inner_mut(&mut self) -> &mut R

Get a mutable reference to the underlying reader.

fn inner(&self) -> &R

Get an immutable reference to the underlying reader.

pub fn read_meta_len(&mut self) -> Result<Option<usize>, ArrowError>

Read the metadata length for the next message from the underlying stream.

Returns

Ok(None) if the the reader signals the end of stream with EOF on the first read

Err(_) if the reader returns an error other than on the first read, or if the metadata length is less than 0.

Ok(Some(_)) with the length otherwise.