arrow_ipc

Module compression

Enums

CompressionCodec

Represents compressing a ipc stream using a particular compression algorithm

Constants

LENGTH_NO_COMPRESSED_DATA 🔒

LENGTH_OF_PREFIX_DATA 🔒

Functions

compress_lz4 🔒

compress_zstd 🔒

decompress_lz4 🔒

decompress_zstd 🔒

read_uncompressed_size 🔒

Get the uncompressed length Notes: LENGTH_NO_COMPRESSED_DATA: indicate that the data that follows is not compressed 0: indicate that there is no data positive number: indicate the uncompressed length for the following data