Module arrow_ipc::convert

Utilities for converting between IPC types and native Arrow types

Structs

FBFieldType 🔒

MessageBuffer

An owned container for a validated Message

Functions

build_field 🔒

Create an IPC Field from an Arrow Field

fb_to_schema

Deserialize a Schema table from flat buffer format to Schema data type

get_data_type 🔒

Get the Arrow data type from the flatbuffer Field table

get_fb_dictionary 🔒

Create an IPC dictionary encoding

get_fb_field_type 🔒

Get the IPC type of a data type

metadata_to_fb

schema_to_fb

Serialize a schema in IPC format

schema_to_fb_offset

try_schema_from_flatbuffer_bytes

Try deserialize flat buffer format bytes into a schema

try_schema_from_ipc_buffer

Try deserialize the IPC format bytes into a schema