arrow_ipc::gen::Schema

Struct Buffer

#[repr(transparent)]
pub struct Buffer(pub [u8; 16]);

A Buffer represents a single contiguous memory segment

Tuple Fields
0: [u8; 16]
Implementations

impl<'a> Buffer

pub fn new(offset: i64, length: i64) -> Self

pub fn offset(&self) -> i64

The relative offset into the shared memory page where the bytes for this buffer starts

pub fn set_offset(&mut self, x: i64)

pub fn length(&self) -> i64

The absolute length (in bytes) of the memory buffer. The memory is found from offset (inclusive) to offset + length (non-inclusive). When building messages using the encapsulated IPC message, padding bytes may be written after a buffer, but such padding bytes do not need to be accounted for in the size here.

pub fn set_length(&mut self, x: i64)