arrow_ipc::gen
Module SparseTensor

Structs

SparseMatrixCompressedAxis

SparseMatrixIndexCSX

Compressed Sparse format, that is matrix-specific.

SparseMatrixIndexCSXArgs

SparseMatrixIndexCSXBuilder

SparseTensor

SparseTensorArgs

SparseTensorBuilder

SparseTensorIndex

SparseTensorIndexCOO

SparseTensorIndexCOOArgs

SparseTensorIndexCOOBuilder

SparseTensorIndexCSF

Compressed Sparse Fiber (CSF) sparse tensor index.

SparseTensorIndexCSFArgs

SparseTensorIndexCSFBuilder

SparseTensorIndexUnionTableOffset

Enums

SparseMatrixIndexCSXOffset

SparseTensorIndexCOOOffset

SparseTensorIndexCSFOffset

SparseTensorOffset

Constants

ENUM_MAX_SPARSE_MATRIX_COMPRESSED_AXISDeprecated

ENUM_MAX_SPARSE_TENSOR_INDEXDeprecated

ENUM_MIN_SPARSE_MATRIX_COMPRESSED_AXISDeprecated

ENUM_MIN_SPARSE_TENSOR_INDEXDeprecated

ENUM_VALUES_SPARSE_MATRIX_COMPRESSED_AXISDeprecated

ENUM_VALUES_SPARSE_TENSOR_INDEXDeprecated

Functions

finish_size_prefixed_sparse_tensor_buffer

finish_sparse_tensor_buffer

root_as_sparse_tensor

Verifies that a buffer of bytes contains a SparseTensor and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_sparse_tensor_unchecked.

root_as_sparse_tensor_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a SparseTensor and returns it.

root_as_sparse_tensor_with_opts

Verifies, with the given options, that a buffer of bytes contains a SparseTensor and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_sparse_tensor_unchecked.

size_prefixed_root_as_sparse_tensor

Verifies that a buffer of bytes contains a size prefixed SparseTensor and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_sparse_tensor_unchecked.

size_prefixed_root_as_sparse_tensor_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a size prefixed SparseTensor and returns it.

size_prefixed_root_as_sparse_tensor_with_opts

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed SparseTensor and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_sparse_tensor_unchecked.