arrow_ipc::gen::Tensor

Struct Tensor

pub struct Tensor<'a> {
    pub _tab: Table<'a>,
}

Fields
_tab: Table<'a>
Implementations

impl<'a> Tensor<'a>

pub const VT_TYPE_TYPE: VOffsetT = 4u16

pub const VT_TYPE_: VOffsetT = 6u16

pub const VT_SHAPE: VOffsetT = 8u16

pub const VT_STRIDES: VOffsetT = 10u16

pub const VT_DATA: VOffsetT = 12u16

pub unsafe fn init_from_table(table: Table<'a>) -> Self

pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr, A: Allocator + 'bldr>( _fbb: &'mut_bldr mut FlatBufferBuilder<'bldr, A>, args: &'args TensorArgs<'args>, ) -> WIPOffset<Tensor<'bldr>>

pub fn type_type(&self) -> Type

pub fn type_(&self) -> Table<'a>

The type of data contained in a value cell. Currently only fixed-width value types are supported, no strings or nested types

pub fn shape(&self) -> Vector<'a, ForwardsUOffset<TensorDim<'a>>>

The dimensions of the tensor, optionally named

pub fn strides(&self) -> Option<Vector<'a, i64>>

Non-negative byte offsets to advance one value cell along each dimension If omitted, default to row-major order (C-like).

pub fn data(&self) -> &'a Buffer

The location and size of the tensorโs data

pub fn type_as_null(&self) -> Option<Null<'a>>

pub fn type_as_int(&self) -> Option<Int<'a>>

pub fn type_as_floating_point(&self) -> Option<FloatingPoint<'a>>

pub fn type_as_binary(&self) -> Option<Binary<'a>>

pub fn type_as_utf_8(&self) -> Option<Utf8<'a>>

pub fn type_as_bool(&self) -> Option<Bool<'a>>

pub fn type_as_decimal(&self) -> Option<Decimal<'a>>

pub fn type_as_date(&self) -> Option<Date<'a>>

pub fn type_as_time(&self) -> Option<Time<'a>>

pub fn type_as_timestamp(&self) -> Option<Timestamp<'a>>

pub fn type_as_interval(&self) -> Option<Interval<'a>>

pub fn type_as_list(&self) -> Option<List<'a>>

pub fn type_as_struct_(&self) -> Option<Struct_<'a>>

pub fn type_as_union(&self) -> Option<Union<'a>>

pub fn type_as_fixed_size_binary(&self) -> Option<FixedSizeBinary<'a>>

pub fn type_as_fixed_size_list(&self) -> Option<FixedSizeList<'a>>

pub fn type_as_map(&self) -> Option<Map<'a>>

pub fn type_as_duration(&self) -> Option<Duration<'a>>

pub fn type_as_large_binary(&self) -> Option<LargeBinary<'a>>

pub fn type_as_large_utf_8(&self) -> Option<LargeUtf8<'a>>

pub fn type_as_large_list(&self) -> Option<LargeList<'a>>

pub fn type_as_run_end_encoded(&self) -> Option<RunEndEncoded<'a>>

pub fn type_as_binary_view(&self) -> Option<BinaryView<'a>>

pub fn type_as_utf_8_view(&self) -> Option<Utf8View<'a>>

pub fn type_as_list_view(&self) -> Option<ListView<'a>>

pub fn type_as_large_list_view(&self) -> Option<LargeListView<'a>>