Module arrow_ipc::reader

Arrow IPC File and Stream Readers

The FileReader and StreamReader have similar interfaces, however the FileReader expects a reader that supports Seeking

Structs

ArrayReader 🔒

State for decoding arrays from an encoded RecordBatch

FileReader

Arrow File reader

StreamReader

Arrow Stream reader

Functions

create_array 🔒

Coordinates reading arrays based on data types.

create_dictionary_array 🔒

Reads the correct number of buffers based on list type and null_count, and creates a list array ref

create_list_array 🔒

Reads the correct number of buffers based on list type and null_count, and creates a list array ref

create_primitive_array 🔒

Reads the correct number of buffers based on data type and null_count, and creates a primitive array ref

read_buffer 🔒

Read a buffer based on offset and length From https://github.com/apache/arrow/blob/6a936c4ff5007045e86f65f1a6b6c3c955ad5103/format/Message.fbs#L58 Each constituent buffer is first compressed with the indicated compressor, and then written with the uncompressed length in the first 8 bytes as a 64-bit little-endian signed integer followed by the compressed buffer bytes (and then padding as required by the protocol). The uncompressed length may be set to -1 to indicate that the data that follows is not compressed, which can be useful for cases where compression does not yield appreciable savings.

read_dictionary

Read the dictionary from the buffer and provided metadata, updating the dictionaries_by_id with the resulting dictionary

read_record_batch

Creates a record batch from binary data using the crate::RecordBatch indexes and the Schema